+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [Breaking] ⬆️ Upgrade dependencies
+- [Breaking] ✨ Update `no-unused-vars` to ignore '_' prefixed variables and arguments
 
 ## 1.12.0 / 2023-06-28